+++ src/main/java/rose/RoseException.java
@@ -1,3 +1,5 @@
+package rose;
+
 public class RoseException extends Exception {
 
     public RoseException(String message) {
