+++ src/main/java/rose/TaskList.java
@@ -1,3 +1,7 @@
+package rose;
+
+import rose.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
