+++ src/main/java/rose/command/ExitCommand.java
@@ -1,4 +1,10 @@
-class ExitCommand extends Command {
+package rose.command;
+
+import rose.Storage;
+import rose.TaskList;
+import rose.Ui;
+
+public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.showClosing();
