+++ src/main/java/rose/command/ListCommand.java
@@ -1,0 +1,12 @@
+package rose.command;
+
+import rose.Storage;
+import rose.TaskList;
+import rose.Ui;
+
+public class ListCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.showTasks(ui);
+    }
+}
