+++ src/main/java/rose/command/UnmarkCommand.java
@@ -1,20 +1,23 @@
+package rose.command;
+
+import rose.Storage;
+import rose.TaskList;
+import rose.Ui;
+
 import java.io.IOException;
 
-class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     private int idx;
 
-    public DeleteCommand(int idx) {
+    public UnmarkCommand(int idx) {
         this.idx = idx;
     }
 
+    @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            Task deleted = tasks.getTask(idx - 1);
-            tasks.deleteTask(idx - 1);
-            ui.showDelete(deleted, tasks.size());
-            //ui.display("Noted. I've removed this task :");
-            //ui.display(deleted.toString());
-            //ui.display(String.format("Now you have %d task in the list.", tasks.size()));
+            tasks.getTask(idx - 1).unmark();
+            ui.showUnmark(tasks.getTask(idx - 1));
         } catch (IndexOutOfBoundsException e) {
             ui.showError("OOPS!!! Task index is out of bounds.");
         }
