+++ src/main/java/rose/task/Event.java
@@ -1,3 +1,5 @@
+package rose.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
