+++ src/main/java/rose/task/TaskType.java
@@ -1,3 +1,5 @@
+package rose.task;
+
 public enum TaskType {
     TODO, EVENT, DEADLINE
 }
